refactor(web-serial-data-access): pi zero prompt 判定を専用サービスへ分離
- PiZeroPromptDetectorService を新設し pi@ シェル / login / password 等の判定を集約
- SerialPromptDetectorService は汎用 matchesPrompt のみに削減
- PiZeroSerialBootstrapService の DI を新サービスへ切替
- 既存 spec のインスタンス化と参照を新サービスへ追従更新

Refs #594

Made-with: Cursor
